+++ src/Models/UI/views/livewire/_partials/allowed-sites.blade.php
@@ -1,5 +1,6 @@
 <x-chief::form.fieldset class="w-full space-y-3" rule="allowed_sites">
-    <x-chief::form.label for="allowed_sites" required>Op welke sites wil je de nieuwe pagina tonen?
+    <x-chief::form.label for="allowed_sites" required>
+        Op welke sites wil je de nieuwe pagina tonen?
     </x-chief::form.label>
 
     @foreach (\Thinktomorrow\Chief\Sites\ChiefSites::all() as $site)
@@ -20,7 +21,8 @@
             <div class="flex grow items-start justify-between gap-2">
                 <div class="space-y-2">
                     <p class="font-medium leading-5 text-grey-700">
-                        {{ $site->name }}: <span class="leading-5 text-grey-400">{{ $site->url }}</span>
+                        {{ $site->name }}:
+                        <span class="leading-5 text-grey-400">{{ $site->url }}</span>
                     </p>
                 </div>
             </div>